Temp fix for bug. Still refactoring and working on race issue with HoardCards
