+++ api/info.json
@@ -14,22 +14,16 @@
       "id": 3
     },
     {
-      "gameId": "H5eAV21t9Y",
-      "userId": 1,
-      "gameStateId": 1,
-      "id": 5
-    },
-    {
-      "gameId": "2V9TFEMjiV",
-      "userId": 1,
-      "gameStateId": 1,
-      "id": 6
-    },
-    {
       "gameId": "NmpDUK2glo",
       "userId": 1,
       "gameStateId": 1,
       "id": 7
+    },
+    {
+      "gameId": "7Gkci5ctCp",
+      "userId": 1,
+      "gameStateId": 1,
+      "id": 8
     }
   ],
   "gameStates": [
